requirements/requirements-prod.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-RESTX==1.2.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.9
marshmallow==3.20.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0